requirements.txt
alembic
dateparser
gTTS
mysql-connector-python
numpy
pandas
pyTelegramBotAPI
python-binance
pytz
regex
requests
SQLAlchemy
Unidecode
urllib3
python-dotenv
pandas_datareader
matplotlib
requests